@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { Moment } from 'src/app/Moment';
 
 @Component({
@@ -11,16 +11,16 @@ export class MomentFormComponent implements OnInit {
   @Output() onSubmit = new EventEmitter<Moment>()
   @Input() btnText!: string;
 
-  momentForm!: FormGroup; // Aqui o FormGroup é apenas declarado
+  momentForm!: UntypedFormGroup; // Aqui o FormGroup é apenas declarado
 
   constructor() { }
 
   ngOnInit(): void {
-    this.momentForm = new FormGroup({ // Aqui o FormGroup é inicializado
-      id: new FormControl(''),
-      title: new FormControl('', [Validators.required]),
-      description: new FormControl('', [Validators.required]),
-      image: new FormControl('')
+    this.momentForm = new UntypedFormGroup({ // Aqui o FormGroup é inicializado
+      id: new UntypedFormControl(''),
+      title: new UntypedFormControl('', [Validators.required]),
+      description: new UntypedFormControl('', [Validators.required]),
+      image: new UntypedFormControl('')
     });
   }
 
